admin_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
bcrypt==4.0.1


# fastapi
# uvicorn
# pydantic
# psycopg2-binary
# pydantic[email]
# python-multipart
